+++ docs/changes.rst
@@ -4,6 +4,7 @@ Weblate 5.5
 Not yet released.
 
 **New features**
+* :ref:`addons` Enable project-wide and site-wide addons.
 
 * API improvements